Add files via upload
semantic.py uses semantic similarity to first compare three different words, then to compare words, in a sentence. It then compares a sentence to a list of sentences. For my own interest, I compare how semantic similarity works with the small and medium language models.

movie_recommendation.ipynb reads the .txt file and uses semantic similarity to find the most similar movie. It gives the user a suggestion for the next movie to watch.